Added Array Messages Selected to Delete
