+++ src/redux/actions/chatActions.js
@@ -70,16 +70,16 @@ export const addMessage = (newMessage) => {
    };
 };
 
-export const deleteMessage = (id) => {
+export const deleteMessage = (idMessagesSelected) => {
    return(dispatch, getState) =>{
       return new Promise (async(resolve, reject) =>{
          try {
             console.log("chatActions: deleteMessage"); //! SOLO PARA PRUEBAS
-            const urlRegister = `https://academlo-whats.herokuapp.com/api/v1/messages/${id}`;
-            await fetch(urlRegister, {
-               "messages": [getState().messages[0]._id, id]
-               });
-            const baseURL = `https://academlo-whats.herokuapp.com/api/v1/messages/${getState().messages[0]._id}/messages`;
+            idMessagesSelected.forEach( async (id) => {
+               const urlDelete = `https://academlo-whats.herokuapp.com/api/v1/messages/${id}`;
+               await fetch(urlDelete); 
+            });
+            const baseURL = `https://academlo-whats.herokuapp.com/api/v1/conversations/${getState().messages[0]._id}/messages`;
             const message = await dispatch(fetchMessages(baseURL));
             resolve("Se ha eliminado el mensaje");
          }catch(error){
